[incubator-kie-issues#1206] Further cleanup/reorganization
